Avoid unneeded HTTP request when rewriting URLs.
Calling `source.get` would potentially trigger two HTTP requests to
Pixiv to gather all the information it needs. It needs to make a
request to the HTML page to get the Original flag, and possibly a
request to the API to get the image directory.

So instead we use `source.get_image_url`, which only needs to make one
request to the API to gather all the information it needs to build the
URL.