Add jump to first & last pagination buttons
Most websites follow this convention:

* << and >> jump to the very first and last pages, respectively
* On non last pages, < and > jump to the previous and next pages

Danbooru didn't do this before this commit. This is only somewhat
annoying on desktop, as you can just jump to the last numbered page or
the first numbered page, but on responsive mode you can't see numbered
pages -- you can only see the arrows. This means that if you're in a
thread in responsive mode, you can't jump to the beginning or the end of
the thread.

With this change, the chevrons behave as expected on both responsive and
desktop layouts. I toyed with hiding the "first" and "last" pages, but
that removes context for the start and the end (and just looks wrong to
me).