main menu: remove aliases and implications pages.
Remove links to the aliases and implications pages from the main menu.
Put them in the submenu on the /tags page instead.

These pages are too rarely used to deserve being in the toplevel menu.
They also weren't easy to find for non-Builders. The toplevel menu
shouldn't be dependent on user level.